2025/4/24

+++ Lampyris.Server.Crypto.Common/Sources/Quote/Data/QuoteTickerData.cs
@@ -6,49 +6,49 @@ public class QuoteTickerData
     public string Symbol = "";
 
     // 最新成交价
-    public double Price;
+    public decimal Price;
 
     // 最新成交的数量，0 代表没有成交量
-    public double LastSize;
+    public decimal LastSize;
 
     // 24小时最高价
-    public double High;
+    public decimal High;
 
     // 24小时最低价
-    public double Low;
+    public decimal Low;
 
     // 24小时成交量，以币为单位
-    public double Volumn;
+    public decimal Volumn;
 
     // 24小时成交量，以张为单位
-    public double Currency;
+    public decimal Currency;
 
     // ticker数据产生时间，Unix时间戳的毫秒数格式，如 1597026383085
     public long Timestamp;
 
     // 涨幅
-    public double ChangePerc;
+    public decimal ChangePerc;
 
     // 涨跌额
-    public double Change;
+    public decimal Change;
 
     // 均价
-    public double Avg => Volumn / Currency;
+    public decimal Avg => Volumn / Currency;
 
     // 标记价格
-    public double MarkPrice;
+    public decimal MarkPrice;
 
     // 指数价格
-    public double IndexPrice;
+    public decimal IndexPrice;
 
     // 资金费率
-    public double FundingRate;
+    public decimal FundingRate;
 
     // 下一次资金时间戳
     public long NextFundingTime;
 
     // 涨速
-    public double RiseSpeed;
+    public decimal RiseSpeed;
 
     // 异动标签
     public List<string> Labels;
